Split out jewel sword puzzle on us
