+++ config/splat.us.stno3.yaml
@@ -52,6 +52,7 @@ segments:
       - [0x1044, .data, e_castle_door]
       - [0x1140, .data, e_background_bushes_trees]
       - [0x123C, .data, entrance_stage_entities_3]
+      - [0x1264, .data, e_jewel_sword_puzzle]
       - [0x1358, .data, entrance_stage_entities_2]
       - [0x13A8, data]
       - [0x1ACC, .data, cutscene]
@@ -170,6 +171,7 @@ segments:
       - [0x394F0, c, e_cavern_door]
       - [0x39C44, c, entrance_weights]
       - [0x3A348, c, entrance_stage_entities_3]
+      - [0x3A4C0, c, e_jewel_sword_puzzle]
       - [0x3AEF0, c, e_stairway]
       - [0x3B548, c, entrance_stage_entities_2]
       - [0x3BF38, c, e_heartroom]
