+++ config/splat.us.stnp3.yaml
@@ -47,6 +47,7 @@ segments:
       - [0xF0C, .data, e_castle_door]
       - [0x1008, .data, e_background_bushes_trees]
       - [0x1104, .data, entrance_stage_entities_3]
+      - [0x1120, .data, e_jewel_sword_puzzle]
       - [0x11CC, .data, entrance_stage_entities_2]
       - [0x122C, data]
       - [0x1300, data]
@@ -171,6 +172,7 @@ segments:
       - [0x33D24, c, e_cavern_door]
       - [0x344B4, c, entrance_weights]
       - [0x34BE8, c, entrance_stage_entities_3]
+      - [0x34D60, c, e_jewel_sword_puzzle]
       - [0x35790, c, e_stairway]
       - [0x35DE8, c, entrance_stage_entities_2]
       - [0x35F60, c, e_heartroom]
