+++ src/st/no3/e_jewel_sword_puzzle.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "no3.h"
+
+#include "../e_jewel_sword_puzzle.h"
